feat: Add TestPage component
This commit adds a new TestPage component to the project. The TestPage component is a simple functional component that renders a div with the text "TestPage". This component will be used for testing purposes.

Note: The commit message follows the convention of starting with a type (feat for feature) followed by a colon and a brief description of the change.